Added reset button to crime log form and added a controller to share API logic for blocks.
